Removing AudioPlayerControll component put in AudioPlayerLoading in order to connect the range input wirh play button
