+++ frontend/src/components/AudioPlayer/index.jsx
@@ -1,4 +1,3 @@
-import AudioPlayerControll from "@components/AudioPlayerControll";
 import AudioPlayerLoading from "@components/AudioPlayerLoading";
 import SAudioPlayer from "./style";
 
@@ -6,7 +5,6 @@ export default function AudioPlayer() {
   return (
     <SAudioPlayer>
       <AudioPlayerLoading />
-      <AudioPlayerControll />
     </SAudioPlayer>
   );
 }
