--- frontend/src/components/AudioPlayerControll/style.jsx
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export default styled.div``;
