+++ frontend/src/pages/Stream/style.jsx
@@ -1,3 +1,3 @@
 import styled from "styled-components";
-export default styled.section``;
 
+export default styled.section``;
